Sc-21460: Refactor Login & minor improvement (#743)
