+++ web/beacon-app/src/features/home/components/QuickViewSummary.tsx
@@ -11,7 +11,7 @@ function QuickViewSummary() {
   const [quickViewData, setQuickViewData] = useState<any>(getDefaultHomeStats()); // by default we will show empty values
   const { tenants } = useFetchTenants();
 
-  console.log('[] tenants', tenants);
+  // console.log('[] tenants', tenants);
 
   const { quickView, error } = useFetchTenantQuickView(tenants?.tenants[0]?.id);
