+++ web/beacon-app/src/features/onboarding/components/steps/workspace/index.tsx
@@ -35,7 +35,7 @@ const WorkspaceStep = () => {
         workspace: values.workspace,
       },
     };
-    console.log(requestPayload);
+    // console.log(requestPayload);
     updateProfile(requestPayload);
   };
