+++ web/beacon-app/src/features/onboarding/layout/index.tsx
@@ -8,8 +8,8 @@ type Props = {
 const OnboardingFormLayout = ({ children }: Props) => {
   return (
     <>
-      <Container max={696} centered className="m-10 mt-20 px-4 xl:mt-10 xl:px-28">
-        <div className="border-[1px] border-[#72A2C0] p-10">{children}</div>
+      <Container max={696} centered className="mt-10 flex  px-4 xl:m-10 xl:mt-20 xl:px-28">
+        <div className="border-[1px] border-[#72A2C0] p-10 px-5">{children}</div>
       </Container>
     </>
   );
